+++ src/main/java/uniandes/edu/co/proyecto/Repositorio/OrdenRepository.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import uniandes.edu.co.proyecto.modelo.Orden;
 
 
-public interface OrdenRepository extends JpaRepository<Orden, Integer>{
+public interface OrdenRepository extends JpaRepository<Orden, Long>{
 
     @Query(value = "SELECT * FROM ordenes", nativeQuery=true )
     Collection<Orden> darOrdenes();
@@ -24,19 +24,19 @@ public interface OrdenRepository extends JpaRepository<Orden, Integer>{
     @Transactional
     @Query(value = "INSERT INTO ordenes (id, estado, fecha_estimada, proveedor, bodega_destino) VALUES(paso.nextval, :estado, :fecha_estimada, :proveedor, :bodega_destino)", nativeQuery = true)
     void insertarOrden(@Param("fecha_estimada")Date fecha_estimada, 
-    @Param("proveedor")Integer proveedor, @Param("bodega_destino")Integer bodega_destino);
+    @Param("proveedor")Long proveedor, @Param("bodega_destino")Long bodega_destino);
 
     @Modifying
     @Transactional
     @Query(value = "UPDATE ordenes SET estado=:'anulada', fecha_estimada=:fecha_estimada,fecha_llegada=:fecha_llegada,fecha_creacion=:fecha_creacion,proveedor=:proveedor,bodega_destino=:bodega_destino WHERE id = :id AND estado IS NOT 'entregada' ", nativeQuery = true)
-    void actualizarOrdenAnulada(@Param("id") Integer id, @Param("fecha_estimada")Date fecha_estimada, @Param("fecha_llegada")Date fecha_llegada, @Param("fecha_creacion")Date fecha_creacion, 
-    @Param("proveedor")Integer proveedor, @Param("bodega_destino")Integer bodega_destino);
+    void actualizarOrdenAnulada(@Param("id") Long id, @Param("fecha_estimada")Date fecha_estimada, @Param("fecha_llegada")Date fecha_llegada, @Param("fecha_creacion")Date fecha_creacion, 
+    @Param("proveedor")Long proveedor, @Param("bodega_destino")Long bodega_destino);
 
     @Modifying
     @Transactional
     @Query(value = "UPDATE ordenes SET estado=:'entregada', fecha_estimada=:fecha_estimada,fecha_llegada=:fecha_llegada,fecha_creacion=:fecha_creacion,proveedor=:proveedor,bodega_destino=:bodega_destino WHERE id = :id AND estado IS NOT 'anulada' ", nativeQuery = true)
-    void actualizarOrdenEntregada(@Param("id") Integer id, @Param("fecha_estimada")Date fecha_estimada, @Param("fecha_llegada")Date fecha_llegada, @Param("fecha_creacion")Date fecha_creacion, 
-    @Param("proveedor")Integer proveedor, @Param("bodega_destino")Integer bodega_destino);
+    void actualizarOrdenEntregada(@Param("id") Long id, @Param("fecha_estimada")Date fecha_estimada, @Param("fecha_llegada")Date fecha_llegada, @Param("fecha_creacion")Date fecha_creacion, 
+    @Param("proveedor")Long proveedor, @Param("bodega_destino")Long bodega_destino);
 
 
     @Modifying
