+++ src/main/java/uniandes/edu/co/proyecto/repositorio/AlmacenajeRepository.java
@@ -17,18 +17,18 @@ public interface AlmacenajeRepository extends JpaRepository<Almacenaje, Almacena
 
     
     @Query(value = "SELECT * FROM almacenajes WHERE bodega = :bodega AND producto = :producto", nativeQuery = true)
-    Almacenaje darAlmacenaje(@Param("bodega") Integer bodega, @Param("producto") Integer producto);
+    Almacenaje darAlmacenaje(@Param("bodega") Long bodega, @Param("producto") Long producto);
 
     @Query(value = "SELECT * FROM almacenajes WHERE bodega = :bodega", nativeQuery = true)
-    Collection<Almacenaje> darAlmacenajesPorBodega(@Param("bodega") Integer bodega);
+    Collection<Almacenaje> darAlmacenajesPorBodega(@Param("bodega") Long bodega);
 
     @Query(value = "SELECT * FROM almacenajes WHERE producto = :producto", nativeQuery = true)
-    Collection<Almacenaje> darAlmacenajesPorProducto(@Param("producto") Integer producto);
+    Collection<Almacenaje> darAlmacenajesPorProducto(@Param("producto") Long producto);
 
 
     /*RFC1 */
     @Query(value = "SELECT almacenajes.bodega, productos.volumen*almacenajes.cantidad/almacenajes.capacidad FROM almacenajes INNER JOIN productos ON almacenajes.producto = productos.cod_barras WHERE bodega = :bodega", nativeQuery = true)
-    Collection<Almacenaje> darIndiceOcupacion(@Param("bodega") Integer bodega);
+    Collection<Almacenaje> darIndiceOcupacion(@Param("bodega") Long bodega);
 
 
 
@@ -36,29 +36,29 @@ public interface AlmacenajeRepository extends JpaRepository<Almacenaje, Almacena
     @Modifying
     @Transactional
     @Query(value = "DELETE FROM almacenajes WHERE bodega = :bodega AND producto = :producto", nativeQuery = true)
-    void eliminarAlmacenaje(@Param("bodega") Integer bodega, @Param("producto") Integer producto);
+    void eliminarAlmacenaje(@Param("bodega") Long bodega, @Param("producto") Long producto);
 
     @Modifying
     @Transactional
     @Query(value = "DELETE FROM almacenajes WHERE bodega = :bodega", nativeQuery = true)
-    void eliminarAlmacenajePorBodega(@Param("bodega") Integer bodega);
+    void eliminarAlmacenajePorBodega(@Param("bodega") Long bodega);
 
     @Modifying
     @Transactional
     @Query(value = "DELETE FROM almacenajes WHERE producto = :producto", nativeQuery = true)
-    void eliminarAlmacenajePorProducto(@Param("producto") Integer producto);
+    void eliminarAlmacenajePorProducto(@Param("producto") Long producto);
 
     @Modifying
     @Transactional
     @Query(value = "UPDATE almacenajes SET costo_promedio = :costo_promedio, capacidad = :capacidad, cantidad = :cantidad, nivel_minimo = :nivel_minimo WHERE bodega = :bodega AND producto = :producto", nativeQuery = true)
-    void actualizarAlmacenaje(@Param("bodega") Integer bodega, @Param("producto") Integer producto, @Param("capacidad") Integer capacidad,
-                    @Param("cantidad") Integer cantidad, @Param("costo_promedio") Integer costo_promedio,
-                    @Param("nivel_minimo") Integer nivel_minimo);
+    void actualizarAlmacenaje(@Param("bodega") Long bodega, @Param("producto") Long producto, @Param("capacidad") Long capacidad,
+                    @Param("cantidad") Long cantidad, @Param("costo_promedio") Long costo_promedio,
+                    @Param("nivel_minimo") Long nivel_minimo);
 
     @Modifying
     @Transactional
     @Query(value = "INSERT INTO almacenajes (bodega, producto, capacidad, cantidad, costo_promedio, nivel_minimo) VALUES ( :bodega, :producto, :capacidad, :cantidad, :costo_promedio, :nivel_minimo)", nativeQuery = true)
-    void insertarAlmacenaje(@Param("bodega") Integer bodega, @Param("producto") Integer producto, @Param("capacidad") Integer capacidad,
-        @Param("cantidad") Integer cantidad, @Param("costo_promedio") Integer costo_promedio,
-        @Param("nivel_minimo") Integer nivel_minimo);
+    void insertarAlmacenaje(@Param("bodega") Long bodega, @Param("producto") Long producto, @Param("capacidad") Long capacidad,
+        @Param("cantidad") Long cantidad, @Param("costo_promedio") Long costo_promedio,
+        @Param("nivel_minimo") Long nivel_minimo);
 }
