+++ src/main/java/uniandes/edu/co/proyecto/repositorio/CategoriaRepository.java
@@ -11,13 +11,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import uniandes.edu.co.proyecto.modelo.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Integer>{
+public interface CategoriaRepository extends JpaRepository<Categoria, Long>{
     @Query(value = "SELECT * FROM categorias", nativeQuery = true)
     Collection<Categoria> darCategorias();
 
     
     @Query(value = "SELECT * FROM categorias WHERE codigo = :codigo", nativeQuery = true)
-    Categoria darCategoria(@Param("codigo") Integer codigo);
+    Categoria darCategoria(@Param("codigo") Long codigo);
 
     @Query(value = "SELECT * FROM categorias WHERE nombre = :nombre", nativeQuery = true)
     Collection<Categoria> darCategoriasPorNombre(@Param("nombre") String nombre);
@@ -26,7 +26,7 @@ public interface CategoriaRepository extends JpaRepository<Categoria, Integer>{
     @Modifying
     @Transactional
     @Query(value = "DELETE FROM categorias WHERE codigo = : codigo", nativeQuery = true)
-    void eliminarCategoria(@Param("codigo") Integer codigo);
+    void eliminarCategoria(@Param("codigo") Long codigo);
 
 
     @Modifying
